+++ src/main/java/com/yanchenko/onlineshop/repository/WareRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface WareRepository extends CrudRepository<Ware, Integer> {
+public interface WareRepository extends CrudRepository<Ware, Long> {
 
     Optional<Ware> findByWareName(String name);
